Create ex_opencv.py
I have been looking for an alternative to manually downloading maps from the google api and plotting points on it in opencv and have been unable to find one untill i came across this library, it is easy to convert a pil image to opencv, so that is what i did, and added some things to be able to zoom and plot points. I have not been able to add a buffer to this, but I have never worked with asyncio before and don't understand it at present.
This is also my first ever contribution to github, so please provide feedback.